Log out status of request_publisher.cc
Reason: when this script is waiting for a connection to the
plugin, it does not output to the terminal that it is waiting.
We might think that it is working on generating the map file
but indeed it is still waiting forever. Log out the waiting
status to the terminal will be helpful.